add requirements
The error message suggests that there was a problem building the pygobject package, which is a dependency of your project. Specifically, it seems that the gobject-introspection-1.0 package is missing.

To resolve this issue, you can try the following steps:

Update your requirements.txt file to specify a version of pygobject that is compatible with the Heroku environment. You can try adding the following line to your requirements.txt file:

makefile
Copy code
pygobject==3.40.1
Commit the changes to your requirements.txt file and redeploy your application to Heroku.

If the issue persists, you can try adding an Aptfile to your project root directory with the following content:

Copy code
libgirepository1.0-dev
The Aptfile is used by Heroku to install system-level dependencies during the build process. This will ensure that the necessary package is installed before attempting to build pygobject.

Commit the Aptfile to your repository and redeploy your application to Heroku.

If you continue to encounter issues, you can try reaching out to Heroku support for further assistance. They may be able to provide specific guidance based on your project's configuration and environment.